Restructure aggregation code
