+++ pylana/aggregation/__init__.py
@@ -1,0 +1,7 @@
+"""
+api for aggregation requests
+"""
+
+from pylana.aggregation.core import AggregationAPI
+from pylana.aggregation.groupings import create_grouping
+from pylana.aggregation.metrics import create_metric
