+++ pylana/aggregation/utils.py
@@ -1,0 +1,9 @@
+"""
+Utility functions used for aggregation
+"""
+
+import pandas as pd
+
+
+def _raise(exception):
+    raise exception
