+++ tests/aggregation/test_aggregation.py
@@ -1,5 +1,6 @@
 import json
 import unittest
+
 import pandas as pd
 
 from pylana import create_api
